server/requirements.txt
# Core MCP Server
mcp>=0.3.0
python-dotenv>=1.0.0
pydantic>=1.10.7

# PII Detection
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
transformers>=4.28.1
torch>=2.0.0

# LLM Support
openai>=1.0.0
google-generativeai>=0.3.0

# HTTP Client for MCP communication
aiohttp>=3.8.0
aiofiles>=23.0.0

# Process management and monitoring
psutil>=5.9.0
backoff>=2.2.1

# Configuration management
PyYAML>=6.0

# FastAPI for main server
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0